moved editor buffer
